refactor(cli): extract shared render helpers to src/cli/render.mjs
doctor/run ハンドラ抽出の前提として、printPlan/printExplain/formatJsonOutput 等の
共有 render ヘルパを cli.mjs から src/cli/render.mjs へ文字単位不変で移す。
cli.mjs と後続の doctor/run ハンドラ双方から import できるようにし、
ハンドラ⇔cli.mjs の循環 import を避ける。振る舞い・出力・exit code は不変。
